added licenses, refactored utility class
