+++ src/main/java/io/github/mzmine/modules/visualization/spectra/simplespectra/mirrorspectra/MirrorScanWindowFX.java
@@ -15,6 +15,7 @@
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
  */
+
 package io.github.mzmine.modules.visualization.spectra.simplespectra.mirrorspectra;
 
 import io.github.mzmine.gui.chartbasics.gui.javafx.EChartViewer;
